+++ network/middleware.go
@@ -10,6 +10,7 @@ import (
 	"github.com/onflow/flow-go/model/flow"
 	"github.com/onflow/flow-go/module/component"
 	"github.com/onflow/flow-go/network/channels"
+	"github.com/onflow/flow-go/network/message"
 )
 
 // Middleware represents the middleware layer, which manages the connections to
@@ -32,13 +33,13 @@ type Middleware interface {
 	// Dispatch should be used whenever guaranteed delivery to a specific target is required. Otherwise, Publish is
 	// a more efficient candidate.
 	// All errors returned from this function can be considered benign.
-	SendDirect(msg *OutgoingMessageScope) error
+	SendDirect(msg *message.OutgoingMessageScope) error
 
 	// Publish publishes a message on the channel. It models a distributed broadcast where the message is meant for all or
 	// a many nodes subscribing to the channel. It does not guarantee the delivery though, and operates on a best
 	// effort.
 	// All errors returned from this function can be considered benign.
-	Publish(msg *OutgoingMessageScope) error
+	Publish(msg *message.OutgoingMessageScope) error
 
 	// Subscribe subscribes the middleware to a channel.
 	// No errors are expected during normal operation.
@@ -71,7 +72,7 @@ type Overlay interface {
 	// Identity returns the Identity associated with the given peer ID, if it exists
 	Identity(peer.ID) (*flow.Identity, bool)
 
-	Receive(*IncomingMessageScope) error
+	Receive(*message.IncomingMessageScope) error
 }
 
 // Connection represents an interface to read from & write to a connection.
