+++ network/mocknetwork/message_validator.go
@@ -3,7 +3,8 @@
 package mocknetwork
 
 import (
-	network "github.com/onflow/flow-go/network"
+	"github.com/onflow/flow-go/network/message"
+
 	mock "github.com/stretchr/testify/mock"
 )
 
@@ -13,11 +14,11 @@ type MessageValidator struct {
 }
 
 // Validate provides a mock function with given fields: msg
-func (_m *MessageValidator) Validate(msg network.IncomingMessageScope) bool {
+func (_m *MessageValidator) Validate(msg message.IncomingMessageScope) bool {
 	ret := _m.Called(msg)
 
 	var r0 bool
-	if rf, ok := ret.Get(0).(func(network.IncomingMessageScope) bool); ok {
+	if rf, ok := ret.Get(0).(func(message.IncomingMessageScope) bool); ok {
 		r0 = rf(msg)
 	} else {
 		r0 = ret.Get(0).(bool)
