+++ network/p2p/unicast/ratelimit/bandwidth_rate_limiter_test.go
@@ -6,7 +6,6 @@ import (
 
 	"github.com/onflow/flow-go/model/flow"
 	libp2pmessage "github.com/onflow/flow-go/model/libp2p/message"
-	"github.com/onflow/flow-go/network"
 	"github.com/onflow/flow-go/network/message"
 
 	"github.com/stretchr/testify/require"
@@ -37,7 +36,7 @@ func TestBandWidthRateLimiter_Allow(t *testing.T) {
 		b[i] = byte('X')
 	}
 
-	msg, err := network.NewOutgoingScope(
+	msg, err := message.NewOutgoingScope(
 		flow.IdentifierList{unittest.IdentifierFixture()},
 		channels.TestNetworkChannel,
 		&libp2pmessage.TestMessage{
@@ -90,7 +89,7 @@ func TestBandWidthRateLimiter_IsRateLimited(t *testing.T) {
 
 	require.False(t, bandwidthRateLimiter.IsRateLimited(peerID))
 
-	msg, err := network.NewOutgoingScope(
+	msg, err := message.NewOutgoingScope(
 		flow.IdentifierList{unittest.IdentifierFixture()},
 		channels.TestNetworkChannel,
 		&libp2pmessage.TestMessage{
