+++ network/validator/not_validator.go
@@ -2,6 +2,7 @@ package validator
 
 import (
 	"github.com/onflow/flow-go/network"
+	"github.com/onflow/flow-go/network/message"
 )
 
 var _ network.MessageValidator = (*NotValidator)(nil)
@@ -17,6 +18,6 @@ func NewNotValidator(validator network.MessageValidator) network.MessageValidato
 	}
 }
 
-func (n NotValidator) Validate(msg network.IncomingMessageScope) bool {
+func (n NotValidator) Validate(msg message.IncomingMessageScope) bool {
 	return !n.validator.Validate(msg)
 }
