+++ network/validator/sender_validator.go
@@ -3,6 +3,7 @@ package validator
 import (
 	"github.com/onflow/flow-go/model/flow"
 	"github.com/onflow/flow-go/network"
+	"github.com/onflow/flow-go/network/message"
 )
 
 var _ network.MessageValidator = &SenderValidator{}
@@ -22,7 +23,7 @@ func ValidateSender(sender flow.Identifier) network.MessageValidator {
 }
 
 // Validate returns true if the message origin id is the same as the sender ID.
-func (sv *SenderValidator) Validate(msg network.IncomingMessageScope) bool {
+func (sv *SenderValidator) Validate(msg message.IncomingMessageScope) bool {
 	return sv.sender == msg.OriginId()
 }
