+++ network/validator/target_validator.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/onflow/flow-go/model/flow"
 	"github.com/onflow/flow-go/network"
+	"github.com/onflow/flow-go/network/message"
 	"github.com/onflow/flow-go/utils/logging"
 )
 
@@ -28,7 +29,7 @@ func ValidateTarget(log zerolog.Logger, target flow.Identifier) network.MessageV
 }
 
 // Validate returns true if the message is intended for the given target ID else it returns false
-func (tv *TargetValidator) Validate(msg network.IncomingMessageScope) bool {
+func (tv *TargetValidator) Validate(msg message.IncomingMessageScope) bool {
 	for _, t := range msg.TargetIDs() {
 		if tv.target == t {
 			return true
